BE-0: rename data_user module to infrastructure_data module
