+++ sources/infrastructure/infrastructure_data/src/main/kotlin/photo/world/data/user/entity/profile/DataAlbum.kt
