+++ sources/infrastructure/infrastructure_data/src/main/kotlin/photo/world/data/user/entity/user/BaseUser.kt
