+++ sources/infrastructure/infrastructure_data/src/main/kotlin/photo/world/data/user/repository/spring/SpringDataPostgresqlUserRepository.kt
