+++ sources/infrastructure/infrastructure_data/src/main/kotlin/photo/world/data/user/repository/spring/profile/SpringDataPostgresqlServiceRepository.kt
